TimeFaker/raspi/requirements.txt
requests
gpiozero
rpi-lgpio
apscheduler
bless
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi[standard]
uvicorn
